+++ mathx/geo/geo.go
@@ -6,7 +6,7 @@ import (
 	"github.com/Spencer1O1/powder_space/v2/mathx"
 )
 
-func SegmentIntersectsCircle(start, end, center mathx.Vec2, radius float64) bool {
+func SegmentIntersectsCircle(start, end, center mathx.Vec2, radius float32) bool {
 	d := end.Sub(start)
 	f := start.Sub(center)
 
@@ -31,11 +31,11 @@ func SegmentIntersectsCircle(start, end, center mathx.Vec2, radius float64) bool
 	return (t1 >= 0 && t1 <= 1) || (t2 >= 0 && t2 <= 1)
 }
 
-func PointInCircle(point, center mathx.Vec2, radius float64) bool {
+func PointInCircle(point, center mathx.Vec2, radius float32) bool {
 	r2 := radius * radius
 	return point.Sub(center).MagSq() <= r2
 }
 
-func SphericalRadiusFromVolume(v float64) float64 {
-	return math.Max(math.Cbrt((3*v)/(4*math.Pi)), 0.5)
+func SphericalRadiusFromVolume(v float32) float32 {
+	return float32(math.Max(math.Cbrt((float64(3*v))/(float64(4*math.Pi))), 0.5))
 }
